backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0

# Redis & Caching
redis==5.0.1
celery==5.3.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# MetaGPT Integration
# metagpt==0.6.0

# LLM & AI
openai==1.3.8
langchain==0.1.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.2
aiofiles==23.2.1
pydantic-extra-types==2.3.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2

# Type Hints
typing-extensions==4.9.0
